ajuste endpoint turma
