+++ requests.rest
@@ -12,3 +12,23 @@ Content-Type: application/json
     "turma_id": "",
     "hobbies":["","",""]
 }
+
+###
+POST http://localhost:3003/turma
+Content-Type: application/json
+{
+    "nome": "São Paulo",
+    "modulo": "0"
+}
+
+###
+GET http://localhost:3003/turma
+
+######
+PUT http://localhost:3003/turma/id
+Content-Type: application/json
+
+{
+   
+    "modulo": "0"
+}
